Add: Cleared page at start of protocol. Fix: Rearranged folders. Drop: unneeded python functions file
Why:
1. We want a clear page on the Matrix chat window so that the start of
one protocol does not blend in with the run of the previous protocol

2. I'm considering using static protocols (where the jupyter file already
has a protocol loaded) rather than dynamically creating a jupyter notebook
for each one. This will allow the code to be easily executed in services
like Blender, which do not allow for the iPython libraries needed to
dynamically alter the current jupyter notebook.

Tags: jupyter notebook, matrix, element